backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
structlog==24.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
rank-bm25==0.2.2
pymongo[srv]==4.7.2
langchain==0.2.11
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langchain_huggingface
langchain_mongodb
sentence-transformers==3.0.1
transformers==4.41.2
torch>=2.1.0
reportlab==4.2.0
matplotlib

openai==1.43.0
cohere==5.5.8
sentence-transformers==3.0.1 
httpx==0.27.2
